Fixed bug with relationsinput description
RelationInputDataManager was trying to render an IntlLabel as if it were a string, for whatever reason. When adding custom relation fields to the plugin, any generated description would be rendered as an object, causing a crash.

This commit changes the require prop type to an intl label, as well as formats the intlLable instead of trying to pass it directly to the component.